+++ common_book/src/txtqbmfxs/java/com/intelligent/reader/adapter/SearchSuggestAdapter.java
@@ -25,7 +25,7 @@ public class SearchSuggestAdapter extends BaseAdapter {
 
     private Context mContext;
     private List<SearchCommonBean> mData;
-    private String editInput;
+    private String editInput="";
 
     public SearchSuggestAdapter(Context context, List<SearchCommonBean> mData, String editInput) {
         this.mContext = context;
@@ -80,7 +80,11 @@ public class SearchSuggestAdapter extends BaseAdapter {
             hodler.iv_type.setImageResource(R.drawable.search_transparent);
         }
         String content = bean.getSuggest();
-        String finalInput = AppUtils.deleteAllIllegalChar(editInput);
+        String finalInput="";
+
+        if(editInput!=null){
+            finalInput = AppUtils.deleteAllIllegalChar(editInput);
+        }
 
         content = content.replaceAll(finalInput, "<font color='#87af4c'>" + finalInput + "</font>");
